Update Method Selection and Planning.md
Changed the engineering method from Plan-based to Agile as we divided the team into two parts, Implementation and Documentation, so that we could work on all parts of the assessments simultaneously.

Replaced “This seemed most appropriate as we anticipated no requirement adjustments to manage during the first assessment. 
” from the previous teams document to what it is now. Justified why we built-up on their previous roadmap and how it will benefit us.

Instead of just discussing about game design as the purpose of our weekly meetings, we changed it and added documentation as well, as it was necessary to be up-to-date with both aspects of the assessment as they go hand-in-hand with one another.

Changed the way they assigned tasks and their reasoning for it to what it is now as we decided that if everyone had almost the similar amount of marks for which they were responsible for, each member would work more efficiently and nothing will be left/forgotten till the last minute.