Update Method Selection and Planning.md
Changed the team meeting tool from Zoom to what it is now as in-person meetings allowed us to separate into smaller teams and work together while allowing us to keep real-time progress of the project. Discord calls were more preferred as compared to zoom as it avoided the hassle of having a meeting id and password to join a meeting and the screensharing capability provided on discord was much better than that provided by zoom, according to the groups point of view.

Changed from Trello to GitHub Issues as this was something new that we thought of trying and also the fact that using GitHub issues, we could keep track of all the commits and why something was changed from the previous project, which would later on help us when writing our change report.